fix(R46): unchecked returns, type truncation, calloc safety
HIGH:
- groebner_parallel.c: Fix work_queue_init calloc unchecked (NULL
  dereference on queue operations). Changed to return int error code.
- groebner_parallel.c: Fix simple_poly_create calloc unchecked (NULL
  dereference on term access). Changed to bool with output param.
- groebner_parallel.c: Fix queue allocation loop calloc unchecked
  (NULL dereference in work_queue_init). Added cleanup on failure.
- solver_core.c: Fix int64_t conflicts truncated to int for modulo
  operation (restart_interval check fails when conflicts > INT_MAX)

MEDIUM:
- runtime_monitor.c: Fix uint32_t truncation of size_t alloc/free
  counts (should cast to uint64_t, not uint32_t)
- solver.c: Fix strlen() result implicitly narrowed to int for buffer
  size calculation (potential negative malloc size)
- bdd_encoding.c: Fix size_t to int truncation for CNF buffer
  remaining tracking
- plugin_system.c: Fix size_t to int truncation for JSON buffer
  remaining tracking
- engine.c: Check engine_restore_frozen_point return value on rollback
- interop.c: Check engine_restore_frozen_point return value on
  command failure rollback

+++ core/src/layer4_reasoning/bdd_encoding.c
@@ -1023,7 +1023,7 @@ bool bdd_to_cnf(BDDNode *bdd, char **out_cnf) {
     }
 
     int offset = 0;
-    int remaining = (int) buf_size;
+    size_t remaining = buf_size;
     int clause_count = 0;
 
     /* DIMACS 头部（先写占位，后面回填） */
